flask-api/requirements.txt
flask
pyjwt
flask-cors
ollama
openai
waitress
requests
cryptography